Keep empty security object (#24)
* Keep empty security object

* Improved openapi-format success output

+++ openapi-format.js
@@ -502,7 +502,7 @@ function openapiFilter(oaObj, options) {
         }
 
         // Remove empty objects
-        if (node && Object.keys(node).length === 0 && node.constructor === Object) {
+        if (node && Object.keys(node).length === 0 && node.constructor === Object && this.parent.key !== 'security') {
             // debugFilterStep = 'Filter - Remove empty objects'
             this.delete();
         }
